+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Update styling of secondary CTAs on `InformationMessageCard`
 
 ## [6.0.3] - 2021-12-21Z
 ### Changed